requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.28
Flask-WTF==1.2.1
Werkzeug==3.0.1
python-dotenv==1.0.1
itsdangerous>=2.1.2